feat: Update input state to null
Reason: setCurrentPage(0) in LocationInput's useEffect is overriding
the setCurrentPage in Controller. This issue can be observed when the
user is on the details job page and flips back. The current page that
they were on would be set back to 0 insted of the value held in local
local storage. By setting the initial input state to null instead of
string data type, this will prevent useEffect from running on mount with
the help of control flow statement